+++ crates/vm/levm/src/vm.rs
@@ -11,26 +11,18 @@ use crate::{
         InternalError, OpcodeSuccess, OutOfGasError, ResultReason, TransactionReport, TxResult,
         TxValidationError, VMError,
     },
-    gas_cost::{
-        self, fake_exponential, ACCESS_LIST_ADDRESS_COST, ACCESS_LIST_STORAGE_KEY_COST,
-        BLOB_GAS_PER_BLOB, CODE_DEPOSIT_COST, COLD_ADDRESS_ACCESS_COST, CREATE_BASE_COST,
-        STANDARD_TOKEN_COST, TOTAL_COST_FLOOR_PER_TOKEN, WARM_ADDRESS_ACCESS_COST,
-    },
+    gas_cost::{self, CODE_DEPOSIT_COST, STANDARD_TOKEN_COST, TOTAL_COST_FLOOR_PER_TOKEN},
     opcodes::Opcode,
     precompiles::{
         execute_precompile, is_precompile, SIZE_PRECOMPILES_CANCUN, SIZE_PRECOMPILES_PRAGUE,
         SIZE_PRECOMPILES_PRE_CANCUN,
     },
+    utils::*,
     AccountInfo, TransientStorage,
 };
 use bytes::Bytes;
 use ethrex_core::{types::TxKind, Address, H256, U256};
-use ethrex_rlp;
-use ethrex_rlp::encode::RLPEncode;
-use keccak_hash::keccak;
-use libsecp256k1::{Message, RecoveryId, Signature};
 use revm_primitives::SpecId;
-use sha3::{Digest, Keccak256};
 use std::{
     cmp::max,
     collections::{HashMap, HashSet},
@@ -66,31 +58,9 @@ pub struct VM {
     pub authorization_list: Option<AuthorizationList>,
 }
 
-pub fn address_to_word(address: Address) -> U256 {
-    // This unwrap can't panic, as Address are 20 bytes long and U256 use 32 bytes
-    let mut word = [0u8; 32];
+pub type AccessList = Vec<(Address, Vec<H256>)>;
 
-    for (word_byte, address_byte) in word.iter_mut().skip(12).zip(address.as_bytes().iter()) {
-        *word_byte = *address_byte;
-    }
-
-    U256::from_big_endian(&word)
-}
-
-pub fn word_to_address(word: U256) -> Address {
-    Address::from_slice(&word.to_big_endian()[12..])
-}
-
-// Taken from cmd/ef_tests/ethrex/types.rs, didn't want to fight dependencies yet
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub struct AccessListItem {
-    pub address: Address,
-    pub storage_keys: Vec<H256>,
-}
-
-type AccessList = Vec<(Address, Vec<H256>)>;
-
-type AuthorizationList = Vec<AuthorizationTuple>;
+pub type AuthorizationList = Vec<AuthorizationTuple>;
 // TODO: We have to implement this in ethrex_core
 #[derive(Debug, Clone, Default, Copy)]
 pub struct AuthorizationTuple {
@@ -100,40 +70,6 @@ pub struct AuthorizationTuple {
     pub v: U256,
     pub r_signature: U256,
     pub s_signature: U256,
-}
-
-pub fn get_valid_jump_destinations(code: &Bytes) -> Result<HashSet<usize>, VMError> {
-    let mut valid_jump_destinations = HashSet::new();
-    let mut pc = 0;
-
-    while let Some(&opcode_number) = code.get(pc) {
-        let current_opcode = Opcode::from(opcode_number);
-
-        if current_opcode == Opcode::JUMPDEST {
-            // If current opcode is jumpdest, add it to valid destinations set
-            valid_jump_destinations.insert(pc);
-        } else if (Opcode::PUSH1..=Opcode::PUSH32).contains(&current_opcode) {
-            // If current opcode is push, skip as many positions as the size of the push
-            let size_to_push =
-                opcode_number
-                    .checked_sub(u8::from(Opcode::PUSH1))
-                    .ok_or(VMError::Internal(
-                        InternalError::ArithmeticOperationUnderflow,
-                    ))?;
-            let skip_length = usize::from(size_to_push.checked_add(1).ok_or(VMError::Internal(
-                InternalError::ArithmeticOperationOverflow,
-            ))?);
-            pc = pc.checked_add(skip_length).ok_or(VMError::Internal(
-                InternalError::ArithmeticOperationOverflow, // to fail, pc should be at least usize max - 31
-            ))?;
-        }
-
-        pc = pc.checked_add(1).ok_or(VMError::Internal(
-            InternalError::ArithmeticOperationOverflow, // to fail, code len should be more than usize max
-        ))?;
-    }
-
-    Ok(valid_jump_destinations)
 }
 
 impl VM {
@@ -228,10 +164,8 @@ impl VM {
                 // CREATE tx
 
                 let sender_nonce = get_account(&mut cache, &db, env.origin).info.nonce;
-                let new_contract_address = VM::calculate_create_address(env.origin, sender_nonce)
-                    .map_err(|_| {
-                    VMError::Internal(InternalError::CouldNotComputeCreateAddress)
-                })?;
+                let new_contract_address = calculate_create_address(env.origin, sender_nonce)
+                    .map_err(|_| VMError::Internal(InternalError::CouldNotComputeCreateAddress))?;
 
                 default_touched_accounts.insert(new_contract_address);
 
@@ -479,7 +413,12 @@ impl VM {
                         match validate_create {
                             Ok(new_address) => {
                                 // Set bytecode to new account if success
-                                self.update_account_bytecode(new_address, contract_code)?;
+                                update_account_bytecode(
+                                    &mut self.cache,
+                                    &self.db,
+                                    new_address,
+                                    contract_code,
+                                )?;
                             }
                             Err(error) => {
                                 // Revert if error
@@ -568,84 +507,16 @@ impl VM {
         matches!(self.tx_kind, TxKind::Create)
     }
 
-    fn get_intrinsic_gas(&self, initial_call_frame: &CallFrame) -> Result<u64, VMError> {
-        // Intrinsic Gas = Calldata cost + Create cost + Base cost + Access list cost
-        let mut intrinsic_gas: u64 = 0;
-
-        // Calldata Cost
-        // 4 gas for each zero byte in the transaction data 16 gas for each non-zero byte in the transaction.
-        let calldata_cost = gas_cost::tx_calldata(&initial_call_frame.calldata, self.env.spec_id)
-            .map_err(VMError::OutOfGas)?;
-
-        intrinsic_gas = intrinsic_gas
-            .checked_add(calldata_cost)
-            .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-
-        // Base Cost
-        intrinsic_gas = intrinsic_gas
-            .checked_add(TX_BASE_COST)
-            .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-
-        // Create Cost
-        if self.is_create() {
-            intrinsic_gas = intrinsic_gas
-                .checked_add(CREATE_BASE_COST)
-                .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-
-            let number_of_words = initial_call_frame.calldata.len().div_ceil(WORD_SIZE);
-            let double_number_of_words: u64 = number_of_words
-                .checked_mul(2)
-                .ok_or(OutOfGasError::ConsumedGasOverflow)?
-                .try_into()
-                .map_err(|_| VMError::Internal(InternalError::ConversionError))?;
-
-            intrinsic_gas = intrinsic_gas
-                .checked_add(double_number_of_words)
-                .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-        }
-
-        // Access List Cost
-        let mut access_lists_cost: u64 = 0;
-        for (_, keys) in self.access_list.clone() {
-            access_lists_cost = access_lists_cost
-                .checked_add(ACCESS_LIST_ADDRESS_COST)
-                .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-            for _ in keys {
-                access_lists_cost = access_lists_cost
-                    .checked_add(ACCESS_LIST_STORAGE_KEY_COST)
-                    .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-            }
-        }
-
-        intrinsic_gas = intrinsic_gas
-            .checked_add(access_lists_cost)
-            .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-
-        // Authorization List Cost
-        // `unwrap_or_default` will return an empty vec when the `authorization_list` field is None.
-        // If the vec is empty, the len will be 0, thus the authorization_list_cost is 0.
-        let amount_of_auth_tuples: u64 = self
-            .authorization_list
-            .clone()
-            .unwrap_or_default()
-            .len()
-            .try_into()
-            .map_err(|_| VMError::Internal(InternalError::ConversionError))?;
-        let authorization_list_cost = PER_EMPTY_ACCOUNT_COST
-            .checked_mul(amount_of_auth_tuples)
-            .ok_or(VMError::Internal(InternalError::GasOverflow))?;
-
-        intrinsic_gas = intrinsic_gas
-            .checked_add(authorization_list_cost)
-            .ok_or(OutOfGasError::ConsumedGasOverflow)?;
-
-        Ok(intrinsic_gas)
-    }
-
     fn add_intrinsic_gas(&mut self, initial_call_frame: &mut CallFrame) -> Result<(), VMError> {
         // Intrinsic gas is the gas consumed by the transaction before the execution of the opcodes. Section 6.2 in the Yellow Paper.
 
-        let intrinsic_gas = self.get_intrinsic_gas(initial_call_frame)?;
+        let intrinsic_gas = get_intrinsic_gas(
+            self.is_create(),
+            self.env.spec_id,
+            &self.access_list,
+            &self.authorization_list,
+            initial_call_frame,
+        )?;
 
         self.increase_consumed_gas(initial_call_frame, intrinsic_gas)
             .map_err(|_| TxValidationError::IntrinsicGasTooLow)?;
@@ -691,60 +562,6 @@ impl VM {
         }
     }
 
-    /// Gets the max blob gas cost for a transaction that a user is willing to pay.
-    fn get_max_blob_gas_price(&self) -> Result<U256, VMError> {
-        let blobhash_amount: u64 = self
-            .env
-            .tx_blob_hashes
-            .len()
-            .try_into()
-            .map_err(|_| VMError::Internal(InternalError::ConversionError))?;
-
-        let blob_gas_used: u64 = blobhash_amount
-            .checked_mul(BLOB_GAS_PER_BLOB)
-            .unwrap_or_default();
-
-        let max_blob_gas_cost = self
-            .env
-            .tx_max_fee_per_blob_gas
-            .unwrap_or_default()
-            .checked_mul(blob_gas_used.into())
-            .ok_or(InternalError::UndefinedState(1))?;
-
-        Ok(max_blob_gas_cost)
-    }
-
-    /// Gets the actual blob gas cost.
-    fn get_blob_gas_price(&self) -> Result<U256, VMError> {
-        let blobhash_amount: u64 = self
-            .env
-            .tx_blob_hashes
-            .len()
-            .try_into()
-            .map_err(|_| VMError::Internal(InternalError::ConversionError))?;
-
-        let blob_gas_price: u64 = blobhash_amount
-            .checked_mul(BLOB_GAS_PER_BLOB)
-            .unwrap_or_default();
-
-        let base_fee_per_blob_gas = self.get_base_fee_per_blob_gas()?;
-
-        let blob_gas_price: U256 = blob_gas_price.into();
-        let blob_fee: U256 = blob_gas_price
-            .checked_mul(base_fee_per_blob_gas)
-            .ok_or(VMError::Internal(InternalError::UndefinedState(1)))?;
-
-        Ok(blob_fee)
-    }
-
-    pub fn get_base_fee_per_blob_gas(&self) -> Result<U256, VMError> {
-        fake_exponential(
-            MIN_BASE_FEE_PER_BLOB_GAS,
-            self.env.block_excess_blob_gas.unwrap_or_default(),
-            get_blob_base_fee_update_fraction_value(self.env.spec_id),
-        )
-    }
-
     /// ## Description
     /// This method performs validations and returns an error if any of the validations fail.
     /// It also makes pre-execution changes:
@@ -755,11 +572,17 @@ impl VM {
     ///   See 'docs' for more information about validations.
     fn prepare_execution(&mut self, initial_call_frame: &mut CallFrame) -> Result<(), VMError> {
         let sender_address = self.env.origin;
-        let sender_account = self.get_account(sender_address);
+        let sender_account = get_account(&mut self.cache, &self.db, sender_address);
 
         if self.env.spec_id >= SpecId::PRAGUE {
             // check for gas limit is grater or equal than the minimum required
-            let intrinsic_gas: u64 = self.get_intrinsic_gas(initial_call_frame)?;
+            let intrinsic_gas: u64 = get_intrinsic_gas(
+                self.is_create(),
+                self.env.spec_id,
+                &self.access_list,
+                &self.authorization_list,
+                initial_call_frame,
+            )?;
 
             // calldata_cost = tokens_in_calldata * 4
             let calldata_cost: u64 =
@@ -799,7 +622,10 @@ impl VM {
 
         // blob gas cost = max fee per blob gas * blob gas used
         // https://eips.ethereum.org/EIPS/eip-4844
-        let max_blob_gas_cost = self.get_max_blob_gas_price()?;
+        let max_blob_gas_cost = get_max_blob_gas_price(
+            self.env.tx_blob_hashes.clone(),
+            self.env.tx_max_fee_per_blob_gas,
+        )?;
 
         // For the transaction to be valid the sender account has to have a balance >= gas_price * gas_limit + value if tx is type 0 and 1
         // balance >= max_fee_per_gas * gas_limit + value + blob_gas_cost if tx is type 2 or 3
@@ -827,11 +653,17 @@ impl VM {
             ));
         }
 
-        let blob_gas_cost = self.get_blob_gas_price()?;
+        let blob_gas_cost = get_blob_gas_price(
+            self.env.tx_blob_hashes.clone(),
+            self.env.block_excess_blob_gas,
+            self.env.spec_id,
+        )?;
 
         // (2) INSUFFICIENT_MAX_FEE_PER_BLOB_GAS
         if let Some(tx_max_fee_per_blob_gas) = self.env.tx_max_fee_per_blob_gas {
-            if tx_max_fee_per_blob_gas < self.get_base_fee_per_blob_gas()? {
+            if tx_max_fee_per_blob_gas
+                < get_base_fee_per_blob_gas(self.env.block_excess_blob_gas, self.env.spec_id)?
+            {
                 return Err(VMError::TxValidation(
                     TxValidationError::InsufficientMaxFeePerBlobGas,
                 ));
@@ -854,7 +686,7 @@ impl VM {
         // technically, the sender will not be able to pay it.
 
         // (3) INSUFFICIENT_ACCOUNT_FUNDS
-        self.decrease_account_balance(sender_address, up_front_cost)
+        decrease_account_balance(&mut self.cache, &mut self.db, sender_address, up_front_cost)
             .map_err(|_| TxValidationError::InsufficientAccountFunds)?;
 
         // (4) INSUFFICIENT_MAX_FEE_PER_GAS
@@ -880,7 +712,7 @@ impl VM {
         self.add_intrinsic_gas(initial_call_frame)?;
 
         // (7) NONCE_IS_MAX
-        self.increment_account_nonce(sender_address)
+        increment_account_nonce(&mut self.cache, &self.db, sender_address)
             .map_err(|_| VMError::TxValidation(TxValidationError::NonceIsMax))?;
 
         // (8) PRIORITY_GREATER_THAN_MAX_FEE_PER_GAS
@@ -983,7 +815,12 @@ impl VM {
         } else {
             // Transfer value to receiver
             // It's here to avoid storing the "to" address in the cache before eip7702_set_access_code() step 7).
-            self.increase_account_balance(initial_call_frame.to, initial_call_frame.msg_value)?;
+            increase_account_balance(
+                &mut self.cache,
+                &mut self.db,
+                initial_call_frame.to,
+                initial_call_frame.msg_value,
+            )?;
         }
         Ok(())
     }
@@ -1016,13 +853,23 @@ impl VM {
                 // If transaction execution results in failure (any
                 // exceptional condition or code reverting), setting
                 // delegation designations is not rolled back.
-                self.decrease_account_balance(receiver_address, initial_call_frame.msg_value)?;
+                decrease_account_balance(
+                    &mut self.cache,
+                    &mut self.db,
+                    receiver_address,
+                    initial_call_frame.msg_value,
+                )?;
             } else {
                 // We remove the receiver account from the cache, like nothing changed in it's state.
                 remove_account(&mut self.cache, &receiver_address);
             }
 
-            self.increase_account_balance(sender_address, initial_call_frame.msg_value)?;
+            increase_account_balance(
+                &mut self.cache,
+                &mut self.db,
+                sender_address,
+                initial_call_frame.msg_value,
+            )?;
         }
 
         // 2. Return unused gas + gas refunds to the sender.
@@ -1047,7 +894,12 @@ impl VM {
             .checked_mul(U256::from(gas_to_return))
             .ok_or(VMError::Internal(InternalError::UndefinedState(1)))?;
 
-        self.increase_account_balance(sender_address, wei_return_amount)?;
+        increase_account_balance(
+            &mut self.cache,
+            &mut self.db,
+            sender_address,
+            wei_return_amount,
+        )?;
 
         // 3. Pay coinbase fee
         let coinbase_address = self.env.coinbase;
@@ -1066,14 +918,19 @@ impl VM {
             .ok_or(VMError::BalanceOverflow)?;
 
         if coinbase_fee != U256::zero() {
-            self.increase_account_balance(coinbase_address, coinbase_fee)?;
+            increase_account_balance(
+                &mut self.cache,
+                &mut self.db,
+                coinbase_address,
+                coinbase_fee,
+            )?;
         };
 
         // 4. Destruct addresses in selfdestruct set.
         // In Cancun the only addresses destroyed are contracts created in this transaction
         let selfdestruct_set = self.accrued_substate.selfdestruct_set.clone();
         for address in selfdestruct_set {
-            let account_to_remove = self.get_account_mut(address)?;
+            let account_to_remove = get_account_mut_vm(&mut self.cache, &self.db, address)?;
             *account_to_remove = Account::default();
         }
 
@@ -1092,7 +949,7 @@ impl VM {
         //  Add created contract to cache, reverting transaction if the address is already occupied
         if self.is_create() {
             let new_contract_address = initial_call_frame.to;
-            let new_account = self.get_account(new_contract_address);
+            let new_account = get_account(&mut self.cache, &self.db, new_contract_address);
 
             let value = initial_call_frame.msg_value;
             let balance = new_account
@@ -1125,54 +982,6 @@ impl VM {
         self.call_frames.last_mut().ok_or(VMError::Internal(
             InternalError::CouldNotAccessLastCallframe,
         ))
-    }
-
-    /// Calculates the address of a new conctract using the CREATE opcode as follow
-    ///
-    /// address = keccak256(rlp([sender_address,sender_nonce]))[12:]
-    pub fn calculate_create_address(
-        sender_address: Address,
-        sender_nonce: u64,
-    ) -> Result<Address, VMError> {
-        let mut encoded = Vec::new();
-        (sender_address, sender_nonce).encode(&mut encoded);
-        let mut hasher = Keccak256::new();
-        hasher.update(encoded);
-        Ok(Address::from_slice(hasher.finalize().get(12..).ok_or(
-            VMError::Internal(InternalError::CouldNotComputeCreateAddress),
-        )?))
-    }
-
-    /// Calculates the address of a new contract using the CREATE2 opcode as follow
-    ///
-    /// initialization_code = memory[offset:offset+size]
-    ///
-    /// address = keccak256(0xff + sender_address + salt + keccak256(initialization_code))[12:]
-    ///
-    pub fn calculate_create2_address(
-        sender_address: Address,
-        initialization_code: &Bytes,
-        salt: U256,
-    ) -> Result<Address, VMError> {
-        let init_code_hash = keccak(initialization_code);
-
-        let generated_address = Address::from_slice(
-            keccak(
-                [
-                    &[0xff],
-                    sender_address.as_bytes(),
-                    &salt.to_big_endian(),
-                    init_code_hash.as_bytes(),
-                ]
-                .concat(),
-            )
-            .as_bytes()
-            .get(12..)
-            .ok_or(VMError::Internal(
-                InternalError::CouldNotComputeCreate2Address,
-            ))?,
-        );
-        Ok(generated_address)
     }
 
     /// Increases gas consumption of CallFrame and Environment, returning an error if the callframe gas limit is reached.
@@ -1249,68 +1058,10 @@ impl VM {
 
         // When updating account storage of an account that's not yet cached we need to store the StorageSlot in the account
         // Note: We end up caching the account because it is the most straightforward way of doing it.
-        let account = self.get_account_mut(address)?;
+        let account = get_account_mut_vm(&mut self.cache, &self.db, address)?;
         account.storage.insert(key, storage_slot.clone());
 
         Ok((storage_slot, storage_slot_was_cold))
-    }
-
-    pub fn increase_account_balance(
-        &mut self,
-        address: Address,
-        increase: U256,
-    ) -> Result<(), VMError> {
-        let account = self.get_account_mut(address)?;
-        account.info.balance = account
-            .info
-            .balance
-            .checked_add(increase)
-            .ok_or(VMError::BalanceOverflow)?;
-        Ok(())
-    }
-
-    pub fn decrease_account_balance(
-        &mut self,
-        address: Address,
-        decrease: U256,
-    ) -> Result<(), VMError> {
-        let account = self.get_account_mut(address)?;
-        account.info.balance = account
-            .info
-            .balance
-            .checked_sub(decrease)
-            .ok_or(VMError::BalanceUnderflow)?;
-        Ok(())
-    }
-
-    pub fn increment_account_nonce(&mut self, address: Address) -> Result<u64, VMError> {
-        let account = self.get_account_mut(address)?;
-        account.info.nonce = account
-            .info
-            .nonce
-            .checked_add(1)
-            .ok_or(VMError::NonceOverflow)?;
-        Ok(account.info.nonce)
-    }
-
-    pub fn decrement_account_nonce(&mut self, address: Address) -> Result<(), VMError> {
-        let account = self.get_account_mut(address)?;
-        account.info.nonce = account
-            .info
-            .nonce
-            .checked_sub(1)
-            .ok_or(VMError::NonceUnderflow)?;
-        Ok(())
-    }
-
-    pub fn update_account_bytecode(
-        &mut self,
-        address: Address,
-        new_bytecode: Bytes,
-    ) -> Result<(), VMError> {
-        let account = self.get_account_mut(address)?;
-        account.info.bytecode = new_bytecode;
-        Ok(())
     }
 
     pub fn update_account_storage(
@@ -1319,7 +1070,7 @@ impl VM {
         key: H256,
         new_value: U256,
     ) -> Result<(), VMError> {
-        let account = self.get_account_mut(address)?;
+        let account = get_account_mut_vm(&mut self.cache, &self.db, address)?;
         let account_original_storage_slot_value = account
             .storage
             .get(&key)
@@ -1330,28 +1081,6 @@ impl VM {
         });
         slot.current_value = new_value;
         Ok(())
-    }
-
-    pub fn get_account_mut(&mut self, address: Address) -> Result<&mut Account, VMError> {
-        if !cache::is_account_cached(&self.cache, &address) {
-            let account_info = self.db.get_account_info(address);
-            let account = Account {
-                info: account_info,
-                storage: HashMap::new(),
-            };
-            cache::insert_account(&mut self.cache, address, account.clone());
-        }
-        cache::get_account_mut(&mut self.cache, &address)
-            .ok_or(VMError::Internal(InternalError::AccountNotFound))
-    }
-
-    /// Gets account, first checking the cache and then the database (caching in the second case)
-    pub fn get_account(&mut self, address: Address) -> Account {
-        get_account(&mut self.cache, &self.db, address)
-    }
-
-    pub fn get_account_no_push_cache(&self, address: Address) -> Account {
-        get_account_no_push_cache(&self.cache, &self.db, address)
     }
 
     fn handle_create_non_empty_account(
@@ -1409,7 +1138,8 @@ impl VM {
             self.accrued_substate
                 .touched_accounts
                 .insert(authority_address);
-            let authority_account_info = self.get_account_no_push_cache(authority_address).info;
+            let authority_account_info =
+                get_account_no_push_cache(&self.cache, &self.db, authority_address).info;
 
             // 5. Verify the code of authority is either empty or already delegated.
             let empty_or_delegated = authority_account_info.bytecode.is_empty()
@@ -1450,8 +1180,8 @@ impl VM {
                 None => {
                     // This is to add the account to the cache
                     // NOTE: Refactor in the future
-                    self.get_account(authority_address);
-                    self.get_account_mut(authority_address)?
+                    get_account(&mut self.cache, &self.db, authority_address);
+                    get_account_mut_vm(&mut self.cache, &self.db, authority_address)?
                 }
             };
 
@@ -1462,7 +1192,7 @@ impl VM {
             };
 
             // 9. Increase the nonce of authority by one.
-            self.increment_account_nonce(authority_address)
+            increment_account_nonce(&mut self.cache, &self.db, authority_address)
                 .map_err(|_| VMError::TxValidation(TxValidationError::NonceIsMax))?;
         }
 
@@ -1480,233 +1210,5 @@ impl VM {
         initial_call_frame.valid_jump_destinations =
             get_valid_jump_destinations(&initial_call_frame.bytecode).unwrap_or_default();
         Ok(refunded_gas)
-    }
-
-    /// Used for the opcodes
-    /// The following reading instructions are impacted:
-    ///      EXTCODESIZE, EXTCODECOPY, EXTCODEHASH
-    /// and the following executing instructions are impacted:
-    ///      CALL, CALLCODE, STATICCALL, DELEGATECALL
-    /// In case a delegation designator points to another designator,
-    /// creating a potential chain or loop of designators,
-    /// clients must retrieve only the first code and then stop following the designator chain.
-    ///
-    /// For example,
-    /// EXTCODESIZE would return 2 (the size of 0xef01) instead of 23 which would represent the delegation designation,
-    /// EXTCODEHASH would return 0xeadcdba66a79ab5dce91622d1d75c8cff5cff0b96944c3bf1072cd08ce018329 (keccak256(0xef01)), and
-    /// CALL would load the code from address and execute it in the context of authority.
-    ///
-    /// The idea of this function comes from ethereum/execution-specs:
-    /// https://github.com/ethereum/execution-specs/blob/951fc43a709b493f27418a8e57d2d6f3608cef84/src/ethereum/prague/vm/eoa_delegation.py#L115
-    pub fn eip7702_get_code(
-        &mut self,
-        address: Address,
-    ) -> Result<(bool, u64, Address, Bytes), VMError> {
-        // Address is the delgated address
-        let account = get_account(&mut self.cache, &self.db, address);
-        let bytecode = account.info.bytecode.clone();
-
-        // If the Address doesn't have a delegation code
-        // return false meaning that is not a delegation
-        // return the same address given
-        // return the bytecode of the given address
-        if !has_delegation(&account.info)? {
-            return Ok((false, 0, address, bytecode));
-        }
-
-        // Here the address has a delegation code
-        // The delegation code has the authorized address
-        let auth_address = get_authorized_address(&account.info)?;
-
-        let access_cost = if self
-            .accrued_substate
-            .touched_accounts
-            .contains(&auth_address)
-        {
-            WARM_ADDRESS_ACCESS_COST
-        } else {
-            self.accrued_substate.touched_accounts.insert(auth_address);
-            COLD_ADDRESS_ACCESS_COST
-        };
-
-        let authorized_bytecode = get_account(&mut self.cache, &self.db, auth_address)
-            .info
-            .bytecode;
-
-        Ok((true, access_cost, auth_address, authorized_bytecode))
-    }
-}
-
-fn get_n_value(op: Opcode, base_opcode: Opcode) -> Result<usize, VMError> {
-    let offset = (usize::from(op))
-        .checked_sub(usize::from(base_opcode))
-        .ok_or(VMError::InvalidOpcode)?
-        .checked_add(1)
-        .ok_or(VMError::InvalidOpcode)?;
-
-    Ok(offset)
-}
-
-fn get_number_of_topics(op: Opcode) -> Result<u8, VMError> {
-    let number_of_topics = (u8::from(op))
-        .checked_sub(u8::from(Opcode::LOG0))
-        .ok_or(VMError::InvalidOpcode)?;
-
-    Ok(number_of_topics)
-}
-
-/// Gets account, first checking the cache and then the database (caching in the second case)
-pub fn get_account(cache: &mut CacheDB, db: &Arc<dyn Database>, address: Address) -> Account {
-    match cache::get_account(cache, &address) {
-        Some(acc) => acc.clone(),
-        None => {
-            let account_info = db.get_account_info(address);
-            let account = Account {
-                info: account_info,
-                storage: HashMap::new(),
-            };
-            cache::insert_account(cache, address, account.clone());
-            account
-        }
-    }
-}
-
-pub fn get_account_no_push_cache(
-    cache: &CacheDB,
-    db: &Arc<dyn Database>,
-    address: Address,
-) -> Account {
-    match cache::get_account(cache, &address) {
-        Some(acc) => acc.clone(),
-        None => {
-            let account_info = db.get_account_info(address);
-            Account {
-                info: account_info,
-                storage: HashMap::new(),
-            }
-        }
-    }
-}
-
-/// Checks if account.info.bytecode has been delegated as the EIP7702 determines.
-pub fn has_delegation(account_info: &AccountInfo) -> Result<bool, VMError> {
-    let mut has_delegation = false;
-    if account_info.has_code() && account_info.bytecode.len() == EIP7702_DELEGATED_CODE_LEN {
-        let first_3_bytes = account_info
-            .bytecode
-            .get(..3)
-            .ok_or(VMError::Internal(InternalError::SlicingError))?;
-
-        if first_3_bytes == SET_CODE_DELEGATION_BYTES {
-            has_delegation = true;
-        }
-    }
-    Ok(has_delegation)
-}
-
-/// Gets the address inside the account.info.bytecode if it has been delegated as the EIP7702 determines.
-pub fn get_authorized_address(account_info: &AccountInfo) -> Result<Address, VMError> {
-    if has_delegation(account_info)? {
-        let address_bytes = account_info
-            .bytecode
-            .get(SET_CODE_DELEGATION_BYTES.len()..)
-            .ok_or(VMError::Internal(InternalError::SlicingError))?;
-        // It shouldn't panic when doing Address::from_slice()
-        // because the length is checked inside the has_delegation() function
-        let address = Address::from_slice(address_bytes);
-        Ok(address)
-    } else {
-        // if we end up here, it means that the address wasn't previously delegated.
-        Err(VMError::Internal(InternalError::AccountNotDelegated))
-    }
-}
-
-fn eip7702_recover_address(auth_tuple: &AuthorizationTuple) -> Result<Option<Address>, VMError> {
-    if auth_tuple.s_signature > *SECP256K1_ORDER_OVER2 || U256::zero() >= auth_tuple.s_signature {
-        return Ok(None);
-    }
-    if auth_tuple.r_signature > *SECP256K1_ORDER || U256::zero() >= auth_tuple.r_signature {
-        return Ok(None);
-    }
-    if auth_tuple.v != U256::one() && auth_tuple.v != U256::zero() {
-        return Ok(None);
-    }
-
-    let rlp_buf = (auth_tuple.chain_id, auth_tuple.address, auth_tuple.nonce).encode_to_vec();
-
-    let mut hasher = Keccak256::new();
-    hasher.update([MAGIC]);
-    hasher.update(rlp_buf);
-    let bytes = &mut hasher.finalize();
-
-    let Ok(message) = Message::parse_slice(bytes) else {
-        return Ok(None);
-    };
-
-    let bytes = [
-        auth_tuple.r_signature.to_big_endian(),
-        auth_tuple.s_signature.to_big_endian(),
-    ]
-    .concat();
-
-    let Ok(signature) = Signature::parse_standard_slice(&bytes) else {
-        return Ok(None);
-    };
-
-    let Ok(recovery_id) = RecoveryId::parse(
-        auth_tuple
-            .v
-            .as_u32()
-            .try_into()
-            .map_err(|_| VMError::Internal(InternalError::ConversionError))?,
-    ) else {
-        return Ok(None);
-    };
-
-    let Ok(authority) = libsecp256k1::recover(&message, &signature, &recovery_id) else {
-        return Ok(None);
-    };
-
-    let public_key = authority.serialize();
-    let mut hasher = Keccak256::new();
-    hasher.update(
-        public_key
-            .get(1..)
-            .ok_or(VMError::Internal(InternalError::SlicingError))?,
-    );
-    let address_hash = hasher.finalize();
-
-    // Get the last 20 bytes of the hash -> Address
-    let authority_address_bytes: [u8; 20] = address_hash
-        .get(12..32)
-        .ok_or(VMError::Internal(InternalError::SlicingError))?
-        .try_into()
-        .map_err(|_| VMError::Internal(InternalError::ConversionError))?;
-    Ok(Some(Address::from_slice(&authority_address_bytes)))
-}
-
-/// After EIP-7691 the maximum number of blob hashes changes. For more
-/// information see
-/// [EIP-7691](https://eips.ethereum.org/EIPS/eip-7691#specification).
-pub const fn max_blobs_per_block(specid: SpecId) -> usize {
-    match specid {
-        SpecId::PRAGUE => MAX_BLOB_COUNT_ELECTRA,
-        SpecId::PRAGUE_EOF => MAX_BLOB_COUNT_ELECTRA,
-        _ => MAX_BLOB_COUNT,
-    }
-}
-
-/// According to EIP-7691
-/// (https://eips.ethereum.org/EIPS/eip-7691#specification):
-///
-/// "These changes imply that get_base_fee_per_blob_gas and
-/// calc_excess_blob_gas functions defined in EIP-4844 use the new
-/// values for the first block of the fork (and for all subsequent
-/// blocks)."
-pub const fn get_blob_base_fee_update_fraction_value(specid: SpecId) -> U256 {
-    match specid {
-        SpecId::PRAGUE => BLOB_BASE_FEE_UPDATE_FRACTION_PRAGUE,
-        SpecId::PRAGUE_EOF => BLOB_BASE_FEE_UPDATE_FRACTION_PRAGUE,
-        _ => BLOB_BASE_FEE_UPDATE_FRACTION,
     }
 }
