+++ crates/vm/levm/tests/tests.rs
@@ -20,8 +20,9 @@ use ethrex_levm::{
         blake2f, bls12_g1msm, ecadd, ecmul, ecpairing, ecrecover, identity, modexp, ripemd_160,
         sha2_256,
     },
-    utils::{new_vm_with_ops, new_vm_with_ops_addr_bal_db, new_vm_with_ops_db, ops_to_bytecode},
-    vm::{word_to_address, Storage, VM},
+    testing::{new_vm_with_ops, new_vm_with_ops_addr_bal_db, new_vm_with_ops_db, ops_to_bytecode},
+    utils::{calculate_create_address, word_to_address},
+    vm::{Storage, VM},
     Environment,
 };
 use std::{borrow::BorrowMut, collections::HashMap, sync::Arc};
@@ -3832,7 +3833,7 @@ fn create_happy_path() {
     let call_frame = vm.current_call_frame_mut().unwrap();
     let returned_address = call_frame.stack.pop().unwrap();
 
-    let expected_address = VM::calculate_create_address(
+    let expected_address = calculate_create_address(
         executing_contract_address,
         executing_contract_before.info.nonce,
     )
